Bug 1127804 - use proper Get(U)int32/64Property*() accessors for the members in mailnews/db/msgdb/src/nsDBFolderInfo.cpp. r=rkent a=Ratty
--HG--
extra : amend_source : 47f7234c851386cea6bada9775ef119ac263e19d

+++ mailnews/local/src/nsMsgBrkMBoxStore.cpp
@@ -217,7 +217,7 @@ NS_IMETHODIMP nsMsgBrkMBoxStore::IsSummaryFileValid(nsIMsgFolder *aFolder,
   nsCOMPtr<nsIDBFolderInfo> folderInfo;
   rv = aDB->GetDBFolderInfo(getter_AddRefs(folderInfo));
   NS_ENSURE_SUCCESS(rv, rv);
-  uint64_t folderSize;
+  int64_t folderSize;
   uint32_t folderDate;
   int32_t numUnreadMessages;
 
@@ -231,7 +231,7 @@ NS_IMETHODIMP nsMsgBrkMBoxStore::IsSummaryFileValid(nsIMsgFolder *aFolder,
   uint32_t actualFolderTimeStamp = 0;
   GetMailboxModProperties(aFolder, &fileSize, &actualFolderTimeStamp);
 
-  if ((int64_t)folderSize == fileSize && numUnreadMessages >= 0)
+  if (folderSize == fileSize && numUnreadMessages >= 0)
   {
     if (!folderSize)
     {
